// types/UserData.ts
export interface UserData {
    password: string;
    days: string;
    months: string;
    years: string;
    firstName: string;
    lastName: string;
    company: string;
    address1: string;
    address2: string;
    country: string;
    state: string;
    city: string;
    zipcode: string;
    mobileNumber: string;
}